move findLayerById to Tools
